ci: fix gcc conversion issue
In file included from /tmp/project/pikuma-2d-game-physics-programming/src/game.c:9:
/tmp/project/pikuma-2d-game-physics-programming/src/physics.c: In function 'CollisionDetect':
/tmp/project/pikuma-2d-game-physics-programming/src/physics.c:559:21: error: conversion from 'int' to 'u8' {aka 'unsigned char'} may change value [-Werror=conversion]
  559 |         u8 Bindex = (vertexIndex + 1) % vertexCount;

+++ src/physics.c
@@ -555,7 +555,7 @@ CollisionDetect(struct entity *a, struct entity *b, contact *contact)
       // find the edge closest to the origin
       for (u8 vertexIndex = 0; vertexIndex < vertexCount; vertexIndex++) {
         u8 Aindex = vertexIndex;
-        u8 Bindex = (vertexIndex + 1) % vertexCount;
+        u8 Bindex = (u8)((vertexIndex + 1) % vertexCount);
         v2 A = polytope[Aindex];
         v2 B = polytope[Bindex];
         v2 AB = v2_sub(B, A);
